update test cases

+++ examples/acousticwave/acousticwave_grow.py
@@ -12,7 +12,11 @@ assemble, interpolate, Expression, Function, TestFunction, dx
 NN = np.array((25, 50, 100, 200))
 ERROR = []
 
-tf = 0.2  # Final time
+# Medium ppties:
+lam = 1.0
+rho = 1.0
+c = np.sqrt(lam/rho)
+tf = 0.2/c  # Final time
 exact_expr = Expression(\
 '(pow(t,2)-(pow(x[0]-.5,2)+pow(x[1]-.5,2)))*(sqrt(pow(x[0]-.5,2)+pow(x[1]-.5,2))<=t)', \
 t=tf)
@@ -25,12 +29,12 @@ for Nxy in NN:
     mesh = UnitSquareMesh(Nxy, Nxy, "crossed")
     q = 1   # Polynomial order
     V = FunctionSpace(mesh, 'Lagrange', q)
-    Dt = h/(q*5.)
+    Dt = h/(q*5.*c)
 
     Wave = AcousticWave({'V':V, 'Vl':V, 'Vr':V})
     Wave.verbose = True
     Wave.exact = interpolate(exact_expr, V)
-    Wave.update({'lambda':1.0, 'rho':1.0, 't0':0.0, 'tf':tf, 'Dt':Dt,\
+    Wave.update({'lambda':lam, 'rho':rho, 't0':0.0, 'tf':tf, 'Dt':Dt,\
     'u0init':Function(V), 'utinit':Function(V)})
     test = TestFunction(V)
     def srcterm(tt):
@@ -41,6 +45,7 @@ for Nxy in NN:
     sol, error = Wave.solve()
     ERROR.append(error)
     print 'relative error = {:.5e}'.format(error)
+
 # Convergence order:
 CONVORDER = []
 for ii in range(len(ERROR)-1):
@@ -48,20 +53,25 @@ for ii in range(len(ERROR)-1):
 print '\n\norder of convergence:', CONVORDER
 
 # Save plots:
-filename, ext = splitext(sys.argv[0])
-if isdir(filename + '/'):   rmtree(filename + '/')
-myplot = PlotFenics(filename)
-myplot.set_varname('p')
-plotp = Function(V)
-for index, pp in enumerate(sol):
-    plotp.vector()[:] = pp[0]
-    myplot.plot_vtk(plotp, index)
-myplot.gather_vtkplots()
+try:
+    boolplot = int(sys.argv[1])
+except:
+    boolplot = 0
+if boolplot > 0:
+    filename, ext = splitext(sys.argv[0])
+    if isdir(filename + '/'):   rmtree(filename + '/')
+    myplot = PlotFenics(filename)
+    myplot.set_varname('p')
+    plotp = Function(V)
+    for index, pp in enumerate(sol):
+        plotp.vector()[:] = pp[0]
+        myplot.plot_vtk(plotp, index)
+    myplot.gather_vtkplots()
 
-fig = plt.figure()
-ax = fig.add_subplot(111)
-ax.loglog(1./NN, ERROR, '-o')
-ax.set_xlabel('h')
-ax.set_ylabel('error')
-fig.savefig(filename + '/convergence.eps')
+    fig = plt.figure()
+    ax = fig.add_subplot(111)
+    ax.loglog(1./NN, ERROR, '-o')
+    ax.set_xlabel('h')
+    ax.set_ylabel('error')
+    fig.savefig(filename + '/convergence.eps')
