+++ examples/acousticwave/acousticwave_source.py
@@ -14,8 +14,11 @@ ERROR = []
 
 tf = 0.048  # Final time
 direction = 1
-exact_expr = Expression('x[i]*(1-x[i])*t', i=direction, t=tf)
-utinit_expr = Expression('x[i]*(1-x[i])', i=direction)
+rho = 1.0   # note: if rho != 1.0, you need to modify rhs accordingly
+lam = 4.0
+c2 = lam/rho
+exact_expr = Expression('x[i]*(1-x[i])*t/c', i=direction, t=tf, c=c2)
+utinit_expr = Expression('x[i]*(1-x[i])/c', i=direction, c=c2)
 # Boundary conditions:
 def u0_boundary(x, on_boundary):
     return (x[direction] < 1e-16 or x[direction] > 1.0-1e-16) and on_boundary
@@ -27,13 +30,12 @@ for Nxy in NN:
     mesh = UnitSquareMesh(Nxy, Nxy, "crossed")
     q = 1   # This example is solved 'exactly' for q>=2
     V = FunctionSpace(mesh, 'Lagrange', q)
-    Dt = h/(q*5.)
+    Dt = h/(q*5.*np.sqrt(c2))
 
     Wave = AcousticWave({'V':V, 'Vl':V, 'Vr':V})
-    Wave.verbose = True
     Wave.exact = interpolate(exact_expr, V)
     Wave.bc = DirichletBC(V, ubc, u0_boundary)
-    Wave.update({'lambda':1.0, 'rho':1.0, 't0':0.0, 'tf':tf, 'Dt':Dt,\
+    Wave.update({'lambda':lam, 'rho':rho, 't0':0.0, 'tf':tf, 'Dt':Dt,\
     'u0init':Function(V), 'utinit':interpolate(utinit_expr, V)})
     test = TestFunction(V)
     source = assemble(Constant('2')*test*dx)
@@ -49,20 +51,25 @@ for ii in range(len(ERROR)-1):
 print '\n\norder of convergence:', CONVORDER
 
 # Save plots:
-filename, ext = splitext(sys.argv[0])
-if isdir(filename + '/'):   rmtree(filename + '/')
-myplot = PlotFenics(filename)
-myplot.set_varname('p')
-plotp = Function(V)
-for index, pp in enumerate(sol):
-    plotp.vector()[:] = pp[0]
-    myplot.plot_vtk(plotp, index)
-myplot.gather_vtkplots()
+try:
+    boolplot = int(sys.argv[1])
+except:
+    boolplot = 0
+if boolplot > 0:
+    filename, ext = splitext(sys.argv[0])
+    if isdir(filename + '/'):   rmtree(filename + '/')
+    myplot = PlotFenics(filename)
+    myplot.set_varname('p')
+    plotp = Function(V)
+    for index, pp in enumerate(sol):
+        plotp.vector()[:] = pp[0]
+        myplot.plot_vtk(plotp, index)
+    myplot.gather_vtkplots()
 
-fig = plt.figure()
-ax = fig.add_subplot(111)
-ax.loglog(1./NN, ERROR, '-o')
-ax.set_xlabel('h')
-ax.set_ylabel('error')
-fig.savefig(filename + '/convergence.eps')
+    fig = plt.figure()
+    ax = fig.add_subplot(111)
+    ax.loglog(1./NN, ERROR, '-o')
+    ax.set_xlabel('h')
+    ax.set_ylabel('error')
+    fig.savefig(filename + '/convergence.eps')
