+++ examples/acousticwave/acousticwave_static.py
@@ -12,7 +12,7 @@ interpolate, Expression, Function
 NN = np.array((25, 50, 100, 200))
 ERROR = []
 
-tf = 1./(4*np.sqrt(2))  # Final time
+tf = 1./(8.*np.sqrt(2))  # Final time
 u0_expr = Expression('sin(2*pi*x[0])*sin(2*pi*x[1])')
 def u0_boundary(x, on_boundary):
     return on_boundary
@@ -24,13 +24,12 @@ for Nxy in NN:
     mesh = UnitSquareMesh(Nxy, Nxy, "crossed")
     q = 1   # Polynomial order
     V = FunctionSpace(mesh, 'Lagrange', q)
-    Dt = h/(q*4.*np.sqrt(2))
+    Dt = h/(q*8.*np.sqrt(2))
 
     Wave = AcousticWave({'V':V, 'Vl':V, 'Vr':V})
-    Wave.verbose = True
     Wave.exact = Function(V)
     Wave.bc = DirichletBC(V, ubc, u0_boundary)
-    Wave.update({'lambda':1.0, 'rho':1.0, 't0':0.0, 'tf':tf, 'Dt':Dt,\
+    Wave.update({'lambda':8.0, 'rho':2.0, 't0':0.0, 'tf':tf, 'Dt':Dt,\
     'u0init':interpolate(u0_expr, V), 'utinit':Function(V)})
     sol, error = Wave.solve()
     ERROR.append(error)
@@ -41,19 +40,25 @@ for ii in range(len(ERROR)-1):
 print '\n\norder of convergence:', CONVORDER
 
 # Save plots:
-filename, ext = splitext(sys.argv[0])
-if isdir(filename + '/'):   rmtree(filename + '/')
-myplot = PlotFenics(filename)
-myplot.set_varname('p')
-plotp = Function(V)
-for index, pp in enumerate(sol):
-    plotp.vector()[:] = pp[0]
-    myplot.plot_vtk(plotp, index)
-myplot.gather_vtkplots()
+try:
+    boolplot = int(sys.argv[1])
+except:
+    boolplot = 0
+if boolplot > 0:
+    filename, ext = splitext(sys.argv[0])
+    if isdir(filename + '/'):   rmtree(filename + '/')
+    myplot = PlotFenics(filename)
+    myplot.set_varname('p')
+    plotp = Function(V)
+    for index, pp in enumerate(sol):
+        plotp.vector()[:] = pp[0]
+        myplot.plot_vtk(plotp, index)
+    myplot.gather_vtkplots()
 
-fig = plt.figure()
-ax = fig.add_subplot(111)
-ax.loglog(1./NN, ERROR, '-o')
-ax.set_xlabel('h')
-ax.set_ylabel('error')
-fig.savefig(filename + '/convergence.eps')
+    fig = plt.figure()
+    ax = fig.add_subplot(111)
+    ax.loglog(1./NN, ERROR, '-o')
+    ax.set_xlabel('h')
+    ax.set_ylabel('error')
+    fig.savefig(filename + '/convergence.eps')
+
