+++ examples/acousticwave/acousticwave_traveling.py
@@ -12,7 +12,11 @@ interpolate, Expression, Function
 NN = np.array((25, 50, 100, 200))
 ERROR = []
 
-tf = 1. # Final time
+# Medium ppties:
+lam = 8.0
+rho = 2.0
+c = np.sqrt(lam/rho)
+tf = 1./c # Final time
 direction = 0
 u0_expr = \
 Expression('100*pow(x[i]-.25,2)*pow(x[i]-0.75,2)*(x[i]<=0.75)*(x[i]>=0.25)', \
@@ -21,7 +25,7 @@ uex_expr = \
 Expression('-100*pow(x[i]-.25,2)*pow(x[i]-0.75,2)*(x[i]<=0.75)*(x[i]>=0.25)', \
 i=direction)
 def u0_boundary(x, on_boundary):
-    return (x[0] < 1e-16 or x[0] > 1.0-1e-16) and on_boundary
+    return (x[direction] < 1e-16 or x[direction] > 1.0-1e-16) and on_boundary
 ubc = Constant("0.0")
 
 for Nxy in NN:
@@ -30,13 +34,12 @@ for Nxy in NN:
     mesh = UnitSquareMesh(Nxy, Nxy, "crossed")
     q = 1   # Polynomial order
     V = FunctionSpace(mesh, 'Lagrange', q)
-    Dt = h/(q*5.)
+    Dt = h/(q*5.*c)
 
     Wave = AcousticWave({'V':V, 'Vl':V, 'Vr':V})
-    Wave.verbose = True
     Wave.exact = interpolate(uex_expr, V)
     Wave.bc = DirichletBC(V, ubc, u0_boundary)
-    Wave.update({'lambda':1.0, 'rho':1.0, 't0':0.0, 'tf':tf, 'Dt':Dt,\
+    Wave.update({'lambda':lam, 'rho':rho, 't0':0.0, 'tf':tf, 'Dt':Dt,\
     'u0init':interpolate(u0_expr, V), 'utinit':Function(V)})
     sol, error = Wave.solve()
     ERROR.append(error)
@@ -47,21 +50,27 @@ CONVORDER = []
 for ii in range(len(ERROR)-1):
     CONVORDER.append(np.log(ERROR[ii+1]/ERROR[ii])/np.log((1./NN[ii+1])/(1./NN[ii])))
 print '\n\norder of convergence:', CONVORDER
-# Save plots:
-filename, ext = splitext(sys.argv[0])
-if isdir(filename + '/'):   rmtree(filename + '/')
-myplot = PlotFenics(filename)
-myplot.set_varname('p')
-plotp = Function(V)
-for index, pp in enumerate(sol):
-    plotp.vector()[:] = pp[0]
-    myplot.plot_vtk(plotp, index)
-myplot.gather_vtkplots()
 
-fig = plt.figure()
-ax = fig.add_subplot(111)
-ax.loglog(1./NN, ERROR, '-o')
-ax.set_xlabel('h')
-ax.set_ylabel('error')
-fig.savefig(filename + '/convergence.eps')
+# Save plots:
+try:
+    boolplot = int(sys.argv[1])
+except:
+    boolplot = 0
+if boolplot > 0:
+    filename, ext = splitext(sys.argv[0])
+    if isdir(filename + '/'):   rmtree(filename + '/')
+    myplot = PlotFenics(filename)
+    myplot.set_varname('p')
+    plotp = Function(V)
+    for index, pp in enumerate(sol):
+        plotp.vector()[:] = pp[0]
+        myplot.plot_vtk(plotp, index)
+    myplot.gather_vtkplots()
+    # convergence plot:
+    fig = plt.figure()
+    ax = fig.add_subplot(111)
+    ax.loglog(1./NN, ERROR, '-o')
+    ax.set_xlabel('h')
+    ax.set_ylabel('error')
+    fig.savefig(filename + '/convergence.eps')
